+++ rio/src/RIO/NonEmpty/Partial.hs
@@ -1,0 +1,11 @@
+module RIO.NonEmpty.Partial
+  (
+  -- * Indexing streams
+    (Data.List.NonEmpty.!!)
+
+  -- * Converting to and from a list
+  , Data.List.NonEmpty.fromList
+
+  ) where
+
+import qualified Data.List.NonEmpty